+++ Extension/src/background/app/app-mv3.ts
@@ -69,6 +69,7 @@ import {
     localeDetect,
     PromoNotificationService,
     filterUpdateService,
+    Telemetry,
 } from '../services';
 import { SettingOption } from '../schema';
 import { getRunInfo } from '../utils';
@@ -311,6 +312,8 @@ export class App {
         // so the event should be dispatched eventually after all initialization
         // is done.
         dispatchEvent(new Event(EXTENSION_INITIALIZED_EVENT));
+
+        await Telemetry.init();
     }
 
     /**