+++ Extension/src/pages/filtering-log/components/FilteringLog/FilteringLog.tsx
@@ -102,7 +102,7 @@ const FilteringLog = observer(() => {
                 NotifierType.CustomFilterAdded,
             ];
 
-            removeListenerCallback = Messenger.createLongLivedConnection(
+            const { onUnload } = Messenger.createLongLivedConnection(
                 Page.FilteringLog,
                 events,
                 async (message: LongLivedConnectionCallbackMessage) => {
@@ -141,6 +141,8 @@ const FilteringLog = observer(() => {
                     }
                 },
             );
+
+            removeListenerCallback = onUnload;
         })();
 
         return () => {