+++ tests/helpers/index.ts
@@ -28,6 +28,7 @@ export {
     getStorageFixturesV10,
     getStorageFixturesV11,
     getStorageFixturesV12,
+    getStorageFixturesV13,
     type StorageData,
 } from './fixtures';